refactor(JsonEditor): move json editor in toolkit folder
