+++ src/Ml.Cli.WebApp/ClientApp/src/Annotation/Toolkit/JsonEditor/Editor.js
