+++ src/Ml.Cli.WebApp/ClientApp/src/Annotation/Toolkit/JsonEditor/EditorContainer.scss
