+++ src/Ml.Cli.WebApp/ClientApp/src/Annotation/Toolkit/JsonEditor/ImagesList.js
